+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - By default all @types modules will be now referenced by TypeScript.
 - Debugger port of headless browser changed from 9222 to 9223.
 
+### Added
+
+- blametest command for randomized searching for Headless chrome bugs.
+ 
 ### Fixed
 
 - Detection of dependencies does not break build in second rebuild. It also does not show as missing dependency when it is in devDependencies. Automatically ignore @types/ dependencies.